+++ pyshacl/constraints/constraint_component.py
@@ -3,11 +3,12 @@
 https://www.w3.org/TR/shacl/#core-components-value-type
 """
 import abc
-from rdflib import BNode, Graph
+from rdflib import BNode
 from pyshacl.consts import *
 import logging
 
 log = logging.getLogger(__name__)
+
 
 class ConstraintComponent(object, metaclass=abc.ABCMeta):
 
@@ -33,7 +34,7 @@ class ConstraintComponent(object, metaclass=abc.ABCMeta):
     def evaluate(self, target_graph, focus_value_nodes):
         return NotImplementedError()
 
-    def make_v_report_description(self, severity, focus_node, value_node=None):
+    def make_v_report_description(self, severity, focus_node, value_node=None, result_path=None):
         constraint = self.shacl_constraint_class()
         constraint_name = self.constraint_name()
         if severity == SH_Violation:
@@ -49,27 +50,36 @@ class ConstraintComponent(object, metaclass=abc.ABCMeta):
             else:
                 val_node_string = str(value_node)
             desc += "\tValue Node: {}\n".format(val_node_string)
-        if self.shape.is_property_shape:
+        if result_path is None and self.shape.is_property_shape:
             result_path = self.shape.path()
-            if result_path is not None:
-                desc += "\tResult Path: {}\n".format(str(result_path))
+        if result_path:
+            desc += "\tResult Path: {}\n".format(str(result_path))
+        for m in self.shape.message:
+            if isinstance(m, rdflib.Literal):
+                desc += "\tMessage: {}\n".format(str(m.value))
+            else:
+                desc += "\tMessage: {}\n".format(str(m))
         return desc
 
-    def make_v_report(self, focus_node, value_node=None):
+    def make_v_report(self, focus_node, value_node=None, result_path=None):
         constraint = self.shacl_constraint_class()
-        severity = self.shape.severity()
-        f_triples = list()
+        severity = self.shape.severity
+        r_triples = list()
         f_node = BNode()
-        f_triples.append((f_node, RDF_type, SH_ValidationResult))
-        f_triples.append((f_node, SH_sourceConstraintComponent, constraint))
-        f_triples.append((f_node, SH_sourceShape, self.shape.node))
-        f_triples.append((f_node, SH_resultSeverity, severity))
-        f_triples.append((f_node, SH_focusNode, focus_node))
+        r_triples.append((f_node, RDF_type, SH_ValidationResult))
+        r_triples.append((f_node, SH_sourceConstraintComponent, constraint))
+        r_triples.append((f_node, SH_sourceShape, self.shape.node))
+        r_triples.append((f_node, SH_resultSeverity, severity))
+        r_triples.append((f_node, SH_focusNode, focus_node))
+        desc = self.make_v_report_description(severity, focus_node, value_node,
+                                              result_path=result_path)
         if value_node:
-            f_triples.append((f_node, SH_value, value_node))
-        if self.shape.is_property_shape:
+            r_triples.append((f_node, SH_value, value_node))
+        if result_path is None and self.shape.is_property_shape:
             result_path = self.shape.path()
-            f_triples.append((f_node, SH_resultPath, result_path))
-        desc = self.make_v_report_description(severity, focus_node, value_node)
+        if result_path:
+            r_triples.append((f_node, SH_resultPath, result_path))
+        for m in self.shape.message:
+            r_triples.append((f_node, SH_resultMessage, m))
         log.info(desc)
-        return desc, f_node, f_triples
+        return desc, f_node, r_triples