+++ pyshacl/inference/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from .custom_rdfs_closure import CustomRDFSSemantics, CustomRDFSOWLRLSemantics